Addressing Comments

+++ components/workflow-mgt/org.wso2.carbon.identity.workflow.impl/src/main/java/org/wso2/carbon/identity/workflow/impl/listener/WorkflowListenerImpl.java
@@ -72,6 +72,9 @@ public class WorkflowListenerImpl implements WorkflowListener {
     @Override
     public void doPreDeleteWorkflow(Workflow workflow) throws WorkflowException {
         WorkflowImplService workflowImplService = WorkflowImplServiceDataHolder.getInstance().getWorkflowImplService();
+        if (workflowImplService == null) {
+            throw new WorkflowException("Error when deleting the Workflow " + workflow.getWorkflowName());
+        }
         workflowImplService.removeBPSPackage(workflow);
     }
